Problemas de migraciones y actualizar estructura de la base de datos

+++ db/migrate/20241004141700_remove_class_id_from_evaluations.rb
@@ -1,7 +1,9 @@
 class RemoveClassIdFromEvaluations < ActiveRecord::Migration[6.1]
   def change
-    remove_foreign_key :evaluations, :courses, column: :class_id, if_exists: true
-    remove_index :evaluations, :class_id, if_exists: true
-    remove_column :evaluations, :class_id
+    if column_exists?(:evaluations, :class_id)
+      remove_foreign_key :evaluations, :courses, column: :class_id, if_exists: true
+      remove_index :evaluations, :class_id, if_exists: true
+      remove_column :evaluations, :class_id
+    end
   end
 end
